Fixed Indexes and Scale
The stacked bar groups no longer extends the stacked bars because everything is now overriden.  The last issue I had was that the last group kept coming out with NaN in its path data.  This was due to the miscalculation of the Scale based on the indexes of the top level data as Stacked Bars does it.  Also, the indexes needed to be calculated off of the number of groups instead of the number of total bars.

Also removed any test code left in.