Fix inconsistent tick rerender
Using an index for a key in RN is not advised, it does not provide a key that is maintainable and does not properly update each node.  It's complicated but there are some very good write-ups on this.  this lead to odd spacing dispite y(value) returning proper values.  The items simply weren't rerendering.